requirements.txt
psycopg2
openai
python-dotenv
flask